+++ web/static/react/assets/ColdStartInterviewPage-DEKcLl3C.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-C9O-DMcj.js";import{t}from"./LegalSkillsProfilePage-eDx-34_c.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as ColdStartInterviewPage};