+++ web/static/react/assets/SkillRunDetailPage-C_DK9XTX.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-C9O-DMcj.js";import{t}from"./LegalSkillsReviewPage-D-TMjnaC.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as SkillRunDetailPage};